refactoring: gostDigest and gostSign are ready
