+++ src/new/gostDigest.js
